+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Re-Enable % value in the enemy tooltips (Requires MDT - https://github.com/Nnoggie/MythicDungeonTools)
 - Re-Enable current pull enemy progress % value (Requires MDT - https://github.com/Nnoggie/MythicDungeonTools)
+- Fixed enemy forces bar value
 
 # MythicPlusTimer 6.0.0